backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
twilio==7.16.0
stripe==5.0.0
pytest==7.2.2
pytest-django==4.5.2